refactor(podcasts): replace polling tool ui with lifecycle card
